Removed the testng dependency, now project is only using junit. Removed unused import statements.
