+++ pom.xml
@@ -22,26 +22,6 @@
             <version>4.19.1</version>
         </dependency>
 
-
-
-        <!-- JUnit Jupiter API -->
-        <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter-api</artifactId>
-            <version>5.10.1</version>
-            <scope>test</scope>
-        </dependency>
-
-
-
-        <!-- TestNG -->
-        <dependency>
-            <groupId>org.testng</groupId>
-            <artifactId>testng</artifactId>
-            <version>7.7.1</version>
-            <scope>test</scope>
-        </dependency>
-
         <!-- ExtentReports -->
         <!-- https://mvnrepository.com/artifact/com.aventstack/extentreports -->
         <dependency>
