+++ src/test/java/packagetests/SauceDemoTest.java
@@ -1,42 +1,17 @@
 package packagetests;
 
-import com.aventstack.extentreports.ExtentReports;
-import com.aventstack.extentreports.ExtentTest;
 import com.aventstack.extentreports.Status;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import org.testng.Assert;
-import org.testng.ITestResult;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
 
 import java.time.Duration;
 
 import static org.openqa.selenium.By.*;
 
-public class SauceDemoTest {
-    WebDriver driver;
-    ExtentReports extent;
-    ExtentTest test;
+public class SauceDemoTest extends BaseTest {
 
-    @BeforeClass
-    public void setup() {
-        extent = ExtentManager.getInstance();
-        //System.setProperty("webdriver.chrome.driver", "path/to/chromedriver");
-    }
-
-    @BeforeMethod
-    public void setupTest() {
-        driver = new ChromeDriver();
-        test = ExtentManager.createTest("SauceDemo Test", "Testing SauceDemo login functionality");
-    }
 
     @Test
     public void testLogin() {
@@ -49,7 +24,7 @@ public class SauceDemoTest {
 
             // Verify login is successful
             boolean isLoggedIn = driver.findElements(cssSelector(".inventory_list")).size() > 0;
-            Assert.assertTrue(isLoggedIn, "Login failed!");
+            Assertions.assertTrue(isLoggedIn, "Login failed!");
 
             test.log(Status.PASS, "Login test executed successfully");
         } catch (Exception e) {
@@ -68,10 +43,10 @@ public class SauceDemoTest {
             WebElement addToCartButton = driver.findElement(xpath("//button[@id='add-to-cart-sauce-labs-backpack']"));
             addToCartButton.click();
 
-            Assert.assertTrue(addToCartButton.isDisplayed(), "addToCartButton is not displayed");
-            Assert.assertTrue(addToCartButton.isEnabled(), "addToCartButton is not enable");
+            Assertions.assertTrue(addToCartButton.isDisplayed(), "addToCartButton is not displayed");
+            Assertions.assertTrue(addToCartButton.isEnabled(), "addToCartButton is not enable");
            boolean isPriceBarDisplayed = driver.findElement(xpath("(//div[@class='pricebar'])[1]")).isDisplayed();
-           Assert.assertTrue(isPriceBarDisplayed, "price bar is not displayed");
+           Assertions.assertTrue(isPriceBarDisplayed, "price bar is not displayed");
 
             test.log(Status.PASS, "product details test executed successfully");
         }
@@ -100,11 +75,4 @@ public class SauceDemoTest {
 //        // Flush after each test method
 //        extent.flush();
 //    }
-
-    @AfterClass
-    public void tearDown() {
-        // Flush at the end of the test class
-        extent.flush();
-        driver.quit();
-    }
 }
